search: factor out job logic in query_converter.go
